*Core: New Sort fixed to compile with MSC
git-svn-id: svn://ultimatepp.org/upp/trunk@5250 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/Algo.h
@@ -832,9 +832,9 @@ void Sort(I l, I h, const Less& less)
 				}
 				break;
 			}
-			IterSwap(l, l + Random(count));     // try some other random elements for median pivot
-			IterSwap(middle, l + Random(count));
-			IterSwap(h - 1, l + Random(count));
+			IterSwap(l, l + (int)Random(count));     // try some other random elements for median pivot
+			IterSwap(middle, l + (int)Random(count));
+			IterSwap(h - 1, l + (int)Random(count));
 			pass++;
 		}
 	}
